Minor refactoring and changes.
Added setMessageDelay() to be 750 ms for ExampleBot to improve responsiveness of the default bot. Merged all Timers in a CardGame into one object called gameTimer. This now serves as the Timer for all TimerTasks for an instance of CardGame and reduces the the number of threads per CardGame to one. Added more documentation. Added a missing entry in blackjack.help.