Fixes for issues #75,#101. New Blackjack feature.
Converted the "debt" property into an "account" property in Player.java. Added the ability for a player to transfer cash to and from their account through the "transfer", "deposit" and "withdraw" commands. Added an extra time penalty for players with debt. Currently it is set to 1 additional minute for each $1000 of debt on top of the existing setting. Added an initialization setting in Blackjack to allow for the dealer to hit on soft 17. By default this is set to false.