Made revisions for issue #106, #109. Minor changes.
Hole cards are now displayed with a showdown message when no one can bet anymore. New instances of PokerPot are now immediately added to the pots ArrayList with currentPot pointing to the last pot in the ArrayList. Changed channel comparisons in ExampleBot to equals() instead of ==. Changed FileLogBot's setMessageDelay() back to 1000.