Made revisions for issue #106. Minor changes.
Hole cards are now displayed once when getNumberCanBet() < 2 and getNumberNotFolded > 1, instead of when all players have gone all in. New instances of PokerPot are now immediately added to the pots ArrayList with currentPot pointing to the last pot in the ArrayList. Changed channel comparisons in ExampleBot to equals() instead of ==. Changed FileLogBot's setMessageDelay() back to 1000.