Added Stats.java to simplify stats manipulation.
Moved all stats manipulation to Stats.java and made required classes inherit from that class. This should greatly simplify adding new stats in the future. Updated all code affected by this change. Added an "fstop" command for blackjack to force a round to end and allowing the bot shutdown commands to work. This is useful when a bug occurs.