Fixed a couple of bugs.
The "get" command now accepts setting names with mixed-case letters. Fixed a bug in Blackjack where if a player leaves during the betting phase, but has already made a bet, would be removed from the game immediately. Such a player is now removed at the end of the round, thus not forfeiting his bet immediately.