Change for issue #91.
Added output of net change for each player at the end of a round of TexasPoker. Changed boolean members to int (0/1) for subclasses of Stats so they could be included in statsMap and thus simplifying the addition of more properties in the future. Added some documentation. All values in players.txt are int except for nick.