Change for issue #102.
Updated the way bets are added to pots for over-bets. A separate side pot is no longer created when it would only contain contributions from one player. Instead, that portion of the bet is returned to the player. The tracking for biggest pot should benefit from this change. The winner of the biggest pot is now also stored.