Added ability to load custom INI files. (issue #93)
Added the ability to load custom INI files when starting Blackjack or TexasPoker by providing the path to custom INI file as a parameter. Added documentation for CasinoBot. Added ability to start CasinoBot with a custom config file by adding it as a run-time argument. Moved class member declarations to the top of the class in CasinoBot, CardGame, TexasPoker and Blackjack.